+++ src/main/java/backend/univfit/domain/apply/application/AnnouncementService.java
@@ -1,0 +1,14 @@
+package backend.univfit.domain.apply.application;
+
+import backend.univfit.domain.apply.api.dto.response.AnnouncementListResponse;
+import backend.univfit.domain.apply.api.dto.response.AnnouncementResponse;
+
+import java.util.List;
+
+public interface AnnouncementService {
+    AnnouncementListResponse getAnnouncementList(List<String> statuses);
+
+    AnnouncementResponse getAnnouncement(Long announcementId);
+
+
+}
